requirements.txt
asgiref==3.2.3
bcrypt==3.1.7
cffi==1.13.2
cryptography==2.8
Django==3.0.3
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.0
paramiko==2.7.1
psycopg2-binary==2.8.4
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
